+++ firebase-firestore/CHANGELOG.md
@@ -2,6 +2,9 @@
 - [feature] Custom objects (POJOs) can now be passed as a field value in
   update(), within `Map<>` objects passed to set(), in array transform
   operations, and in query filters.
+- [feature] DocumentSnapshot.get() now supports retrieving fields as
+  custom objects (POJOs) by passing a Class<T> instance, e.g.
+  `snapshot.get("field", CustomType.class)`.
 
 # 17.1.2
 - [changed] Changed the internal handling for locally updated documents that